View update

+++ Interfaces/IHttpService.cs
@@ -4,7 +4,7 @@ namespace Code_editor.Interfaces
 {
     public interface IHttpService
     {
-        Task<ResponseBody> SendCodeAsync(string language, string sourceCode, int memoryLimit, int timeLimit);
+        Task<ResponseBody> SendCodeAsync(string language, string sourceCode, int memoryLimit, double timeLimit);
       //  Task<string> PollCompilationStatus(string statusUpdateUrl);
     }
 }
